requirements.txt
tensorflow-gpu==1.15.2
numpy==1.17.4
pandas==0.24.2
scipy==1.4.1
scikit_learn==0.22
